FIX clean and refactor tests code
